requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.0
python-decouple>=3.8
psycopg2-binary>=2.9
django-cors-headers>=4.3